relocate missing page objects
