+++ x-pack/platform/test/functional/page_objects/index.ts
@@ -25,6 +25,7 @@ import { SearchSessionsPageProvider } from './search_sessions_management_page';
 import { GraphPageObject } from './graph_page';
 import { MaintenanceWindowsPageProvider } from './maintenance_windows_page';
 import { BannersPageObject } from './banners_page';
+import { NavigationalSearchPageObject } from './navigational_search';
 
 // just like services, PageObjects are defined as a map of
 // names to Providers. Merge in Kibana's or pick specific ones
@@ -49,4 +50,5 @@ export const pageObjects = {
   userProfiles: UserProfilePageProvider,
   maintenanceWindows: MaintenanceWindowsPageProvider,
   banners: BannersPageObject,
+  navigationalSearch: NavigationalSearchPageObject,
 };
