+++ x-pack/solutions/observability/test/functional/page_objects/index.ts
@@ -13,14 +13,18 @@ import { InfraMetricsExplorerProvider } from './infra_metrics_explorer';
 import { InfraSavedViewsProvider } from './infra_saved_views';
 import { UptimePageObject } from './uptime_page';
 import { AssetDetailsProvider } from './asset_details';
+import { ObservabilityPageProvider } from './observability_page';
+import { AlertControlsProvider } from './alert_controls';
 
 export const pageObjects = {
   ...platformPageObjects,
+  alertControls: AlertControlsProvider,
   assetDetails: AssetDetailsProvider,
   infraHome: InfraHomePageProvider,
   infraHostsView: InfraHostsViewProvider,
   infraLogs: InfraLogsPageProvider,
   infraMetricsExplorer: InfraMetricsExplorerProvider,
   infraSavedViews: InfraSavedViewsProvider,
+  observability: ObservabilityPageProvider,
   uptime: UptimePageObject,
 };
