+++ x-pack/test/functional/page_objects/index.ts
@@ -25,6 +25,7 @@ import { SearchSessionsPageProvider } from '@kbn/test-suites-xpack-platform/func
 import { GraphPageObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/graph_page';
 import { MaintenanceWindowsPageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/maintenance_windows_page';
 import { BannersPageObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/banners_page';
+import { NavigationalSearchPageObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/navigational_search';
 import { ApiKeysPageProvider } from './api_keys_page';
 import { CrossClusterReplicationPageProvider } from './cross_cluster_replication_page';
 import { DetectionsPageObject } from '../../security_solution_ftr/page_objects/detections';
@@ -39,11 +40,8 @@ import { InfraLogsPageProvider } from './infra_logs_page';
 import { IngestPipelinesPageProvider } from './ingest_pipelines_page';
 import { LicenseManagementPageProvider } from './license_management_page';
 import { LogstashPageObject } from './logstash_page';
-import { NavigationalSearchPageObject } from './navigational_search';
 import { ObservabilityLogsExplorerPageObject } from './observability_logs_explorer';
 import { DatasetQualityPageObject } from './dataset_quality';
-import { ObservabilityPageProvider } from './observability_page';
-import { AlertControlsProvider } from './alert_controls';
 import { RemoteClustersPageProvider } from './remote_clusters_page';
 import { RollupPageObject } from './rollup_page';
 import { ShareSavedObjectsToSpacePageProvider } from './share_saved_objects_to_space_page';
@@ -90,8 +88,6 @@ export const pageObjects = {
   navigationalSearch: NavigationalSearchPageObject,
   observabilityLogsExplorer: ObservabilityLogsExplorerPageObject,
   datasetQuality: DatasetQualityPageObject,
-  observability: ObservabilityPageProvider,
-  alertControls: AlertControlsProvider,
   remoteClusters: RemoteClustersPageProvider,
   reporting: ReportingPageObject,
   roleMappings: RoleMappingsPageProvider,
